도메인 업데이트

+++ src/main/java/ac/su/kiosk/domain/OrderItem.java
@@ -30,5 +30,5 @@ public class OrderItem {
     int quantity;
 
     @Column(nullable = false)
-    BigDecimal price;
+    long price;
 }
